JAMES-3078 small exception refactoring for jmap-draft
