+++ server/protocols/jmap-draft/src/main/java/org/apache/james/jmap/draft/exceptions/UnauthorizedException.java
@@ -20,4 +20,10 @@
 package org.apache.james.jmap.draft.exceptions;
 
 public class UnauthorizedException extends RuntimeException {
+    public UnauthorizedException() {
+    }
+
+    public UnauthorizedException(Throwable cause) {
+        super(cause);
+    }
 }
